server/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.8.2
jinja2==3.1.4
psycopg2-binary==2.9.9
boto3==1.34.131
httpx==0.27.0
python-jose[cryptography]==3.3.0
weasyprint==62.3
alembic==1.13.2

# AI/ML Dependencies
openai==1.54.3
pillow==10.4.0
pytesseract==0.3.13
textblob==0.18.0
langchain==0.3.0
langchain-openai==0.2.0
chromadb==0.5.15
sentence-transformers==3.0.1

